perf(gateway): isolate publication snapshot from executor imports
