+++ src/gateway/github-publication-git-transport.ts
@@ -1,10 +1,39 @@
 import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
+import { runCommandBuffered } from "../process/exec.js";
 import { githubPublicationUnsafeConfigArgs } from "./github-publication-base.js";
 
 type GitCommandOptions = { cwd?: string; env?: NodeJS.ProcessEnv; input?: string };
 type GitCommandResult = { code: number | null; stdout: Buffer };
+
+export async function runPublicationCommand(argv: string[], options: GitCommandOptions = {}) {
+  return await runCommandBuffered(argv, {
+    ...(options.cwd ? { cwd: options.cwd } : {}),
+    env: {
+      ...(options.env ?? process.env),
+      GIT_NO_REPLACE_OBJECTS: "1",
+      // Pin every command against repository hooks; explicit hook-disabling -c flags stay stronger.
+      GIT_CONFIG_COUNT: "1",
+      GIT_CONFIG_KEY_0: "core.hooksPath",
+      GIT_CONFIG_VALUE_0: os.devNull,
+    },
+    ...(options.input !== undefined ? { input: options.input } : {}),
+    timeoutMs: 60_000,
+    maxOutputBytes: 256 * 1024,
+  });
+}
+
+export async function requirePublicationCommand(
+  argv: string[],
+  options: GitCommandOptions = {},
+): Promise<string> {
+  const result = await runPublicationCommand(argv, options);
+  if (result.code !== 0) {
+    throw new Error(`${argv[0]} command failed`);
+  }
+  return result.stdout.toString("utf8").trim();
+}
 
 export async function assertSafeGitPublicationWorkspace(
   cwd: string,
@@ -76,7 +105,7 @@ async function readOptionalAttributeFile(file: string): Promise<Buffer | undefin
   }
 }
 
-export async function assertGitHubPublicationTreeHasNoFilters(
+async function assertGitHubPublicationTreeHasNoFilters(
   cwd: string,
   workspaceTree: string,
   run: (argv: string[], options?: GitCommandOptions) => Promise<GitCommandResult>,
@@ -136,6 +165,84 @@ export async function assertGitHubPublicationTreeHasNoFilters(
     if (contents) {
       assertNoGitFilterAttributes(contents);
     }
+  }
+}
+
+export async function captureGitHubPublicationWorkspaceSnapshot(params: {
+  cwd: string;
+  assertCurrent?: () => void;
+}): Promise<{ sourceHeadCommit: string; sourceIndexTree: string; workspaceTree: string }> {
+  const step = async <T>(operation: () => Promise<T>): Promise<T> => {
+    params.assertCurrent?.();
+    const result = await operation();
+    params.assertCurrent?.();
+    return result;
+  };
+  await step(() => assertSafeGitPublicationWorkspace(params.cwd, runPublicationCommand));
+  const sourceHeadCommit = await step(
+    async () =>
+      await requirePublicationCommand(["git", "rev-parse", "--verify", "HEAD^{commit}"], {
+        cwd: params.cwd,
+      }),
+  );
+  const sourceIndexTree = await step(
+    async () =>
+      await requirePublicationCommand(
+        ["git", "-c", `core.hooksPath=${os.devNull}`, "-c", "core.fsmonitor=false", "write-tree"],
+        { cwd: params.cwd },
+      ),
+  );
+  const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-github-snapshot-"));
+  try {
+    const env: NodeJS.ProcessEnv = {
+      GIT_ATTR_NOSYSTEM: "1",
+      GIT_CONFIG_GLOBAL: os.devNull,
+      GIT_CONFIG_SYSTEM: os.devNull,
+      GIT_INDEX_FILE: path.join(tempDir, "index"),
+    };
+    await step(async () => {
+      await requirePublicationCommand(
+        [
+          "git",
+          "-c",
+          `core.hooksPath=${os.devNull}`,
+          "-c",
+          "core.fsmonitor=false",
+          "read-tree",
+          sourceHeadCommit,
+        ],
+        { cwd: params.cwd, env },
+      );
+    });
+    await step(async () => {
+      await requirePublicationCommand(
+        [
+          "git",
+          "-c",
+          `core.attributesFile=${os.devNull}`,
+          "-c",
+          `core.hooksPath=${os.devNull}`,
+          "-c",
+          "core.fsmonitor=false",
+          "add",
+          "-A",
+        ],
+        { cwd: params.cwd, env },
+      );
+    });
+    const workspaceTree = await step(
+      async () =>
+        await requirePublicationCommand(
+          ["git", "-c", `core.hooksPath=${os.devNull}`, "-c", "core.fsmonitor=false", "write-tree"],
+          { cwd: params.cwd, env },
+        ),
+    );
+    await step(() =>
+      assertGitHubPublicationTreeHasNoFilters(params.cwd, workspaceTree, runPublicationCommand),
+    );
+    return { sourceHeadCommit, sourceIndexTree, workspaceTree };
+  } finally {
+    await fs.rm(tempDir, { recursive: true, force: true });
   }
 }
